BJS2-15622 - Интеграция календаря с Google Calendar + Пофиксил по комментам
