+++ src/main/java/faang/school/projectservice/controller/CalendarController.java
@@ -7,7 +7,6 @@ import faang.school.projectservice.service.CalendarService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,20 +14,19 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-@Controller
+@RestController
 @RequestMapping("/api/v1/calendar")
 @RequiredArgsConstructor
 public class CalendarController {
 
     private final CalendarService calendarService;
 
-    @PostMapping("/new")
-    @ResponseBody
+    @PostMapping()
     @ResponseStatus(HttpStatus.CREATED)
     public CalendarDto createCalendarForProject(@RequestParam Long projectId,
                                                 @RequestBody @Valid CalendarDto calendarDto) {
@@ -36,8 +34,7 @@ public class CalendarController {
     }
 
 
-    @PostMapping("/{calendarId}/event/new")
-    @ResponseBody
+    @PostMapping("/{calendarId}/event")
     @ResponseStatus(HttpStatus.CREATED)
     public EventDto createEvent(@PathVariable String calendarId,
                                 @RequestBody @Valid EventDto eventDto) {
@@ -46,7 +43,6 @@ public class CalendarController {
 
 
     @GetMapping("/{calendarId}/event/{eventId}")
-    @ResponseBody
     @ResponseStatus(HttpStatus.OK)
     public EventDto getEvent(@PathVariable String calendarId,
                              @PathVariable String eventId) {
@@ -60,16 +56,14 @@ public class CalendarController {
         calendarService.deleteEvent(calendarId, eventId);
     }
 
-    @PostMapping("/{calendarId}/event")
-    @ResponseBody
+    @PostMapping("/{calendarId}/events")
     @ResponseStatus(HttpStatus.CREATED)
     public List<EventDto> createEvents(@PathVariable String calendarId,
                                        @RequestBody List<EventDto> eventDtoList) {
         return calendarService.createEvents(calendarId, eventDtoList);
     }
 
-    @GetMapping("/{calendarId}/event")
-    @ResponseBody
+    @GetMapping("/{calendarId}/events")
     @ResponseStatus(HttpStatus.OK)
     public List<EventDto> getEvents(@PathVariable String calendarId) {
         return calendarService.getEvents(calendarId);
@@ -77,7 +71,6 @@ public class CalendarController {
 
 
     @PostMapping("/{calendarId}/acl")
-    @ResponseBody
     @ResponseStatus(HttpStatus.CREATED)
     public AclRule addAclRule(@PathVariable String calendarId,
                               @RequestParam String email,
@@ -88,7 +81,6 @@ public class CalendarController {
 
 
     @GetMapping("/{calendarId}/acl")
-    @ResponseBody
     @ResponseStatus(HttpStatus.OK)
     public List<AclRule> getAclRules(@PathVariable String calendarId) {
         return calendarService.getAclRules(calendarId);
@@ -96,7 +88,6 @@ public class CalendarController {
 
 
     @DeleteMapping("/{calendarId}/acl/{ruleId}")
-    @ResponseBody
     @ResponseStatus(HttpStatus.OK)
     public void deleteAclRule(@PathVariable String calendarId,
                               @PathVariable String ruleId) {
